pyproject.toml
[project]
name = "tasknote"
version = "0.1.0"
description = "Personal task lists application"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "dj-database-url>=2.3.0",
    "django>=5.1.3",
    "django-compressor>=4.5.1",
    "django-cotton>=1.5.0",
    "django-grappelli>=4.0.1",
    "django-htmx>=1.21.0",
    "django-redis>=5.4.0",
    "django-tailwind[reload]>=3.8.0",
    "factory-boy>=3.3.1",
    "inject>=5.2.1",
    "mysqlclient>=2.2.7",
    "pillow>=11.0.0",
    "pip>=25.0.1",
    "pytest-django>=4.10.0",
    "python-decouple>=3.8",
    "selenium>=4.26.1",
    "unipath>=1.1",
    "uwsgi>=2.0.28",
]

[dependency-groups]
dev = [
    "coverage>=7.6.8",
    "django-stubs[compatible-mypy]>=5.1.3",
    "djlint>=1.36.1",
    "mypy>=1.15.0",
    "ruff>=0.7.2",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py", "testing/python/*.py"]
addopts = ["--reuse-db", "--nomigrations"]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "core.settings"
